Move Rolling into Window and update doc_utils
